virtually_sharded_blobstore: move awaited check to inner struct of Ticket
Summary:
The `awaited` boolean on the `Ticket` structure and corresponding `Drop` makes it impossible to move out of the fields of `Ticket`.  This is inconvenient, especially since it only enables a unit test check.

Move the field to a structure that's internal to the `Ticket`.  We still check for failure to await, as dropping the `Ticket` with a `CheckAwait` that hasn't been marked as awaited will still panic during tests when the inner `CheckAwait` gets dropped.

Reviewed By: Croohand

Differential Revision: D44270321

fbshipit-source-id: 29429c40243b31dcc1a5046969b85fb7a15fd0aa