PrjFS: fix compiler warning and possible bug
Summary:
XLOG(DFATAL) is non-fatal in production builds, so it's possible we'll fall through here. It's unclear what the compiler will do in this scenario, so let's make sure we return folly::unit to avoid any compiler weirdness.

This may be causing a frequent crash on Sandcastle hosts where we're panicking after logging this DFATAL error. The crash looks like it's coming from an invalid instruction, so falling through here could be the root cause.

Reviewed By: xavierd

Differential Revision: D44314543

fbshipit-source-id: 6d23ffa221d7af4e7be641349eab1eb9027443a1