Render draft PRs
Summary:
Now that we support passing `--draft`, we need to render diffs as drafts in the UI.

This diff adds support for showing draft github PRs.

We need to fetch `isDraft` in our graphql. For some reason, this isn't a PR `state`, even though it supercedes the OPEN state. So although we fetch it as `isDraft`, it's much more convenient in our code to treat it as another state.

Reviewed By: muirdm

Differential Revision: D44343542

fbshipit-source-id: 3cd324d44b6345da3e92e9a3f1ffccac3c8921eb