Respect draft checkbox when confirming ghstack vs sl pr
Summary:
Small fix, if you submit for the first time in a github repo, it will ask if you want to use ghstack or sl pr. That submit process doesn't respect the draft checkbox you may have selected in the commit info.

I don't think we need to show the draft checkbox _again_ in the interstitial confirmation, since you already selected it before pressing submit, which is the normal condition for using the draft checkbox.

Reviewed By: muirdm

Differential Revision: D44342933

fbshipit-source-id: c720153dd103b98a7936bf3dc3e76ff608007bce