virtually_sharded_blobstore: ensure future is removed even if cancelled
Summary:
As Croohand pointed out on D44183928, if the shared future is cancelled, the `WeakShared` will linger in the collection until the next read of that key.  If the read of that key never comes, it will never be cleaned up.

Switch to a scopeguard, that we move into the future.  This will execute the code to remove the shared future whether the future succeeds, fails, or is dropped.

Reviewed By: Croohand

Differential Revision: D44342772

fbshipit-source-id: 37fb15d126d8289d61ab33f64f6e1cf9f7a8550d