stats: log streams stats when streams are cancelled
Summary:
When streams are cancelled we lose all of their stats, as they only get logged after the stream is polled to completion.

Make the stream stats callback synchronous (we don't actually use its async nature), which means we can call it from a `Drop` implementation, allowing us to log the stats even when the stream is not completed.

To distinguish the two cases, we add a "completed" flag, which is true if the stream was polled to completion.

This also makes `completion_time` optional, as we don't have a value for this if the stream was never polled at all when it got dropped.

`TimedStream` had an implicit fuse-like nature (once completed the stream would never yield anything again), which we rely upon, so make sure this property is maintained.

Reviewed By: liubov-dmitrieva

Differential Revision: D44331959

fbshipit-source-id: 5a7d9875a4d7fd4276c304b6c30ff45c6b990b38