cmdlib: move tokio runtime out of environment
Summary:
The environment is passed into the async runtime via an `Arc`.  If the last reference to the environment is dropped from within the runtime, then we end up destroying the runtime from within the runtime, which is an error.

Move the runtime out of the environment, and store it in the `MononokeApp` (new) or on the stack in `main` (old).  This means the runtime will live on the stack outside of the async runtime, and so execution of all async code should be complete when they are dropped.

Reviewed By: YousefSalama

Differential Revision: D44299775

fbshipit-source-id: 9a93ba49a8bbf376defdd0b5ddb1d6546b5ff12b