status: fix dir-becomes-symlink detection (non-watchman)
Summary:
Fix detection of a directory being replaced by a symlink. When the walker iterates tracked files not seen during the walk it now directly emits "deleted" events instead of going through the file detector. If tracked files had the same size and mtime "through" the symlink, the file detector's lstat() call would make it seem like there were no changes.

The alternative approach is to "audit" paths before we lstat them, but that is slow and we want to avoid if possible.

Reviewed By: quark-zju

Differential Revision: D44187186

fbshipit-source-id: 5678d6027b74283608bd0430e7acd3ae16cfb046