status: fix dir-becomes-symlink detection (watchman)
Summary:
Use the "exists" field from watchman to detect when a file has been moved behind a symlink. Watchman reports it properly as not-exists, but if we naively vfs.metadata(path), it may appear to be unchanged.

I added another layer of Option<> to the file metadata so watchman fs can explicitly say "this file doesn't exist", as opposed to only "I don't have metadata".

The alternative approach is to "audit" paths before we lstat them, but that is slow and we want to avoid if possible.

Reviewed By: quark-zju

Differential Revision: D44187187

fbshipit-source-id: e24a43cbdb36d9aabf2b0b7ae0c4d2eab66394c4