Make commit tree more dense
Summary:
I got some feedback that the old ISL was more information dense than the new one. This is true for 2 main reasons:
1. more padding
2. showing Diff badges in a second row instead of next to commits

(1) we can improve. I think the new ISL has a bit too much padding. So we can adjust some of these numbers to make it more information dense. This diff attempts to do this in a few ways:
- reduce spacing between commit groups
- reduce spacing between diff badge and the diff title it belongs to
- increase spacing between commits slightly to compensate, so that diff badges are still clearly belonging to their commit
- reduce extra space below commits with no Diff attached (this helps a lot)
- reduce space for stack actions ("submit stack")

We could reduce it further, but we do want some padding. Too much density makes it hard to parse what's going on.

Reason (2) for less density is that DIffs are on separate lines now. I kind of like this change, even though it sacrifices some density, since it makes it a lot more visually consistent.

perhaps if this continues to be an issue we'll add an option to show diff badges next to commits instead of below them in some sort of "dense" mode. But I'd like to avoid that if possible.

Reviewed By: quark-zju

Differential Revision: D44469854

fbshipit-source-id: 8a9437183e0950a80e77c851b3421370cdaef139