copytrace: make read_renamed_metadata API async
Summary:
Our current practice is only using `block_on` in Python bindings, so this diff is to change the `read_rename_metadata` to an Async function and also remove the `block_on`s in its implementations, so it is consistent with other APIs.

In the longterm, we'd like to remove those Async APIs, but we will to use a lib to help us migrate them in an automatic way, consistent APIs will make the migration process eaiser.

Reviewed By: quark-zju

Differential Revision: D44199939

fbshipit-source-id: a75d1356e373f3d883da4ab972520a665587eb23