Remove unnecessary sl CLI args
Summary:
We had a bunch of CLI args leftover from the old ISL. We should clean these up so we don't pass a ton of unnecessary flags that make commands in your history really hard to understand.

- `--config ui.merge=:merge3` can be done with an env variable
- `progress.renderer=none` isn't needed because we use `SL_AUTOMATION=true`
- `fbamend.education=` is an old dead code path

Arguably we could also try to use an env var for `--noninteractive`, but I kind of prefer the flag so you can kind of tell that it's not something human spawned.

Reviewed By: quark-zju

Differential Revision: D44430437

fbshipit-source-id: 0b22a8912ebf6e9ccb608ce791e3045aa01a6790