symlinks: make eagerepo aware of Windows symlink option
Summary: Recently we added the `experimental.windows-symlinks` config to make repos initialized with that config add `windowssymlinks` to the working copy requirements. That, however, didn't apply to the `eagerrepo`. This meant that tests initializing repos using eagerrepo would not have this behavior, which made writing `.t` tests expecting that in the requirements somewhat awkward.

Reviewed By: quark-zju

Differential Revision: D44444534

fbshipit-source-id: def5c30267aa91bf6d2a72514644538285f80b8b