Show copied/renamed in UI
Summary:
Add support for copied/renamed files in the list of changes.

When we fetch data from `sl status`, we now fetch it with `--copies`, so that we can see if a file was copied or renamed.

If a file is copied, it adds a `copy: string` field, with the path it was copied from.

If a file is renamed, it's just like a copy, but the original file is also removed and shows up as another change with status `'R'`.

We pass this data directly to the UI. It gets processed there to be visually displayed as
`oldPath → newPath`.

This involves hiding removed files that are the original, removed paths, and then also resorting files to group copies and renames together, plus adding the `→` label and rename icon, etc.

You could imagine doing such preprocessing on the server instead of the client. However, since most of our sl integration speaks in terms unrelated to the copy, we want to keep around the original changes. This way our optimistic state appliers can deal with all the changes and we can pass paths to operations normally. This makes the copy/rename support almost entirely visual, with little chance to affect commands being run.

The exception to that is that we use the visually displayed files with checkboxes for a few commands, like quick commit, amend, and discard. In those cases, for renames, we actually want to treat clicking the checkbox as if you're selecting both the newly created file, and the original, removed file. This can be done from the selection code and the rest of the code can just treat is as an opaque selection without caring that they're "linked" by being a rename together.

TODO: this doesn't yet work when fetching changes in the commit info view via `sl log`. We need to do the equivalent of `--copies` in another diff for that... However this is currently very expensive to fetch it seems. So we may limit this to uncommitted changes for the mean time.

Reviewed By: muirdm

Differential Revision: D44482086

fbshipit-source-id: 8ab9a6df61382d096d9da03d82f5f10b48011374