remove multiple unnessecary retries within hg importer fetches
Summary:
Currently a single eden fetch causes 13 attempts to fetch the data.

By default hg hs 3 configured retries for edenapi fetch operations.

In the direct rust fetching code we have disabled these retries because we
use the importer as a retry, so these are not worth it.

We should disable them in the importer too.

There are still many retries. The importer internally will attempt 2 retries
and eden retries the importer once when it fails. So there are at least 5
attempts to fetch data still.

Those other retries are perhaps ensuring correctness, so we need to be more
careful with them. But these should be safe to remove.

This is a 3x reduction in retry attempts.

Reviewed By: fanzeyi

Differential Revision: D44482325

fbshipit-source-id: 96a501c8e9421e5d8aef92b697a6ed5e19c69194