move findScopeNameForPath utility into shared textmate-lib folder
Summary:
`findScopeNameForPath()` is a pretty important utility with
respect to our other TextMate logic because it leverages the
metadata we have to pick the appropriate TextMate grammar
to apply to a snippet of code based on the file path. As such,
it is compelling to make it shared with the other files in `textmate-lib`.

Reviewed By: evangrayk

Differential Revision: D44493277

fbshipit-source-id: 2c737f876a73084e360c01f33f5531bfa15ef814