Fix hitting multiple merge conflicts among multiple files
Summary:
I ran into a state where I hit merge conflicts but the UI refused to show me the correct conflicts. I debugged this to find that this happens when you hit conflicts multiple times (e.g. rebase a stack and hit conflicts in several commits), among multiple files that are different in each conflicts.

For example,
- rebase commits A and B onto another commit C
- hit conflicts in file1.txt from commit A
- see those in ISL properly, resolve those, then click continue
- hit conflicts in file2.txt from commit B
- bug: it wouldn't show file2.txt's conflicts in ISL.

This is due to a trick we were doing to keep stable ordering of conflicts. We tried to remember past conflicts, and then re-arrange new conflicts to match. However, we neglected to also include other new conflicts into the set. This meant if you hit multiple conflicts, it would not consider those new conflicts.

The fix is to more properly add all new conflicts into the set. I did some refactoring to make this easier to follow as well

Reviewed By: muirdm

Differential Revision: D44390578

fbshipit-source-id: db1898c0efb145c8c2a6053b822394fc1617c17c