Revert in commit info sidebar
Summary:
Until now we only allowed reverting files from the uncommitted changes list. But it's also useful to revert back to how files were in a different commit. Usually this is useful within your stack, to go backwards in the stack. This gives you uncommitted changes equivalent to undoing the changes since that commit.

To do this, we need to start showing revert buttons next to files in the commit info, and passing the hash that they relate to to the revert funciton. We also need to modify the revert operation to take in an optional hash.

Reviewed By: muirdm

Differential Revision: D44473824

fbshipit-source-id: a7c9fc200886c6ecc5966c24c719193193d71d2e