Really fix test that sometimes breaks contbuild
Summary:
D44297278 was an attempt at fixing this test, but the regex for lines with DEBUG
wasn't actually correct as it's not actually the first pattern in the strings being grepped.
It may be preceded by some spaces.

Yesterday, we saw a contbuild breakage caused by this.
https://www.internalfb.com/intern/sandcastle/job/9007200167370789/insights

Relax the grep step to achieve the intent of the previous diff.

Reviewed By: RajivTS

Differential Revision: D44536438

fbshipit-source-id: a40719f4669b55b2442715bc9350b8dcce7a9747