do not hardcode definition of fetchGrammar() in TextMateGrammarManifest.ts
Summary:
Change the way we generate `TextMateGrammarManifest.ts` so that
the implementation for `fetchGrammar()` is no longer a property on
each `Grammar`. Instead, expose the data needed to fetch the
definition of the grammar via `fileName` and `fileType` properties
on `Grammar` and have the caller use this information as they see fit.

In our case, we move `fetchGrammar()` into our application-specific
`reviewstack/src/textmate/` folder instead of the reusable `textmate-lib/` folder.

The remaining bit of generated code that makes hardcoded assumptions about
the directory structure is this line:

```
import type {Grammar, LanguageConfiguration} from '../../textmate-lib/types';
```

Reviewed By: evangrayk

Differential Revision: D44533571

fbshipit-source-id: fad555d5e65dedcb2bdbd6001660c999590a0542