Add custom circle elipsis icon for signals in progress
Summary:
This is kind of a nit-pick, but I didn't like that we used a normal ellipsis for the "in progress" Diff signals. In other places in the UI, ellipsis means different things, and this being ... is inconsistent wit the other signal icons which are surrounded by a circle.

We did this because Codicons don't have a circle ellipsis icon. So let's just add our own.

Reviewed By: quark-zju

Differential Revision: D44515856

fbshipit-source-id: 0893a5ec7cd23bb1a8001c5af8533720ad96612d