Fix flaky `cross-repo-globalrevs` test
Summary:
This test used to be detected as flaky in mode/dev-rust-oss and flaky, but not detected as such in mode/opt.

This was due to a race condition between fresh and warm bookmarks, which take a while to update due to WBC.

Using the helper functions to wait for bookmarks to get warm, or directly querying fresh ids, fixes the issue.

Reviewed By: RajivTS

Differential Revision: D44543648

fbshipit-source-id: f70b16609fe995341b84a77e590efdc3d71b4fdf