Allow submitting multi-selection as draft
Summary:
If using a code review provider that supports submitting multiple commits, when rendering them in the sidebar, we weren't offering to submit as a draft. This diff adds the draft checkbox to this screen, in the same way as for individual commits.

We also need to pluralize the tooltip text if you're going to submit multiple commits. Luckily we set up the SubmitAsDraftCheckbox to take in multiple commits to submit already.

Reviewed By: quark-zju

Differential Revision: D44560868

fbshipit-source-id: 1bd543b10f8fd6f5f6128384b2a7a25839b6b802