github: run 'pr_marker' as a hook
Summary:
We made 'pr_marker' run inline instead of as hook in D43276199, but this can cause
nested event loop runtime error in `sl pr pull`. This diff makes 'pr_marker' run as a
Python hook to avoid the nested event loop.

The difference between current hook (this diff) and previous hook is that we are running
it as a Python object instead of shelling out to `sl` command.

https://github.com/facebook/sapling/issues/58

Reviewed By: quark-zju

Differential Revision: D44531268

fbshipit-source-id: 5270b7172d7b2c22a9c9e715d7cb5b42181cff53