add hg import queue events to ActivityBuffer
Summary:
This is a slight behavior change, but I need queue events published to
ActivityBuffer so I can track the live set of fetches.

The impact will be that `eden trace hg --retroactive` will show 2/3 of
the number of events that it did. We can bump the capacity if that's
problematic.

Reviewed By: xavierd

Differential Revision: D44594520

fbshipit-source-id: 5b22169403402d729ff929b4b4bec6bb76d54737