Don't use rtl for file path truncation
Summary:
We were previously being fancy and using `text-direction: rtl` (right-to-left) to display file paths. This was combined with `text-overflow: ellipsis`, so that long paths would be truncated on the left side with `...`:
```
...ittedChanges.tsx
```
This way, among a list of many files, you're not seeing the same base path for each of them, but rather the ending of the file name which is what you care about.

However, we made a change to use disambiguated file paths anyway, which means we don't often show full file paths. It only really happens now if you have two files with the same name and subfolders causing long disambiguation, or you have a very short screen width.

Regardless, the fancy rtl system has several major downsides. `rtl` is intended to set the text rendering format, not to control ellipses. This means certain characters are re-arranged in a very confusing way. `.gitignore` becomes `gitignore.` for example.

I found a way to work around this, using `unicode-bidi: bidi-override`. But rtl brings a lot of weird issues. To do this, we'd have to reverse the letters in the label of the text. The `rtl` then un-reverses them. But this means the DOM has reversed characters, and if you tried to select it and copy it you'd copy reversed text. It's bad for accessibility.

We can approximate this feature without the ellipsis just by being a bit more crafty in our CSS, and it works pretty much fine without any of these weird accessibility downsides.

This also fixes the `gitignore.` issue.

Reviewed By: muirdm

Differential Revision: D44641831

fbshipit-source-id: 16dcbab7d4cffee16d0fe16a60ec7adb1dcf1b0d