Fix subscriptions not getting disposed on connection close
Summary:
I noticed a bug where closing ISL and re-opening it would cause the vscode extension to no longer report uncommitted changes. There was a "webview is disposed" error in the console too.

Looks like this was an error introduced by adding more complex subscriptions. We just need to dispose those subscriptions when disposing the entire connection. This connection disposal happens when the webview is closed. Re-opening causes an error to be thrown in the repo code path, so it never sent the message back to the new server, even though non-subscription messages make it through just fine.

Reviewed By: muirdm

Differential Revision: D44693737

fbshipit-source-id: 27a0330f7f9a52bd1a4b8b7d53fe1017814c5e7a