Add timeout to reduce flakey test
Summary: I noticed our CI job that runs `verify-addons-folder` was somewhat flakey, always on the comparison view test. It was timing out when waiting for the comparison view to load and send a diff fetch request to the server. I think this is because of dynamic importing of the comparison view javascript, which seemingly takes time on the CI machines. So just bumping the timeout on this `waitFor` assertion to reduce flakiness.

Reviewed By: zzl0

Differential Revision: D44727954

fbshipit-source-id: de09a8f1e8a186d42a32b903bfa1dab0b1921241