Add env var flag for showing full test output
Summary:
We use React Testing Library for testing our UI states, which is mostly awesome. Sometimes it's annoying if a lot of tests fail, because it spews out the entire DOM that doesn't match a particular selector. If more than 1-2 tests fail, then the test output is truncated and you can't see which tests in particular failed.

There's a way to configure the error messages, which lets you remove this extra context. Here I set this config when an env var is set, so that you can run `HIDE_RTL_DOM_ERRORS=1 yarn test` and get more readable output

Reviewed By: zzl0

Differential Revision: D44725386

fbshipit-source-id: 6f62739f9bfa7417361cc00aa421010e715062bf