remove accidental network fetch
Summary:
This continues my slashing of fetch retries. This diff is -1 fetch retry.

DatapackStore::getTree is not a local only tree access. It may fallback to
fetching from edenAPI.

DatapackStore::getTree is used in the hg importer fallback code path for when
normal edenAPI rust binding fetching fails. It's unnessecary to be doing this
extra network fetch here, because we only want one edenAPI attempt and then to
fall back to the importer.

If I had to guess why this is here:
IIRC this method use to be a local only access to the hgcache. That seems to
line up with the logging comments around this fetch. So I think when the
datapack store gained remote fetching capabilities this one just wasn't changed.

With this change we are now down to what looks like 5 network fetch attempts.

Reviewed By: xavierd

Differential Revision: D44523824

fbshipit-source-id: f5e784660b396ad3a8f1e7effcd7551c3b126521