make NFS Server run on windows
Summary:
I'm getting nfs to run on windows to prototype it and see how feasible it
might be as an option on Windows. PrjFS has a very different model than EdenFS,
and that has made EdenFS correctness on Windows very difficult. NFS may be
easier to get correct, though the performance is suspect. Just exploring
options here.

After this change the NFS Server opens a socket to accept incoming connections!

Reviewed By: xavierd

Differential Revision: D44154505

fbshipit-source-id: 10c679cb390c2bab99591656d6fe96ccdf4443fe