Make parseCommitMessageFields use dynamic config
Summary:
like the previous diff, but for parsing from string into commitMessageFields instead of converting into a string.

This uses the logic we use internally (adapted slightly for generality). Again, this is overkill for OSS, where we just take the entire description. Eventually, this will be possibly useful in OSS where you might set this via config.

Reviewed By: muirdm

Differential Revision: D44688587

fbshipit-source-id: 4b4f178283732bf94c137f7ed58c7d8bfaaf027a