Make emptyCommitMessageFields not part of utils
Summary:
Context:
Start a set of incremental refactors which will make the CommitMessageFieldsUtils type only contain the field schema, and no other utils. From there, we'll make that field config a recoil atom so it can be determiend completely dynamically by the server / sl config.

This is part 1, where we make `emptyCommitMessageFields` not part of the utils object, and instead just a util function which takes in the config.

Reviewed By: muirdm

Differential Revision: D44688585

fbshipit-source-id: d9ed56fa0975514452ff547bdea58aaaed120653