Consolidate "name" and "key" in commit message field config
Summary:
Now that we're moving towards dynamic commit message field configs, we should make it easier to define. The "key" value we had here was purely for convenience in typescript, so we could use things like `commitMessageFields.testPlan`. However, since this is dynamic, we can't really do this anyway, so it makes more sense to just unify everything to use the same label value.

So, if we parse a commit message like:
```
  my commit
  Summary: my summary!
  Test Plan: my test plan!
```
Then we would extract this into an object like:
```
{
  "Title": "my commit",
  "Summary": "my summary!",
  "Test Plan": "my test plan!",
}
```

This makes the field config simpler to define.

We may need to add a new field back into this config, which allows for display-only names, which may be useful internally (e.g. we could show "Diff" instead of "Differential Revision").

Reviewed By: muirdm

Differential Revision: D44689896

fbshipit-source-id: c95afd83e06f444cd735302524b589f3cba6a305