Make commitMessageFieldsToString use dynamic config
Summary: Continue refactoring as in previous diff, now for conversion from commitMessageFields back into a string. This is much more complicated than necessary for OSS right now, because OSS only uses the title and entire description. But soon we'll support arbitrary fields for OSS too, so they'll need this as well.

Reviewed By: muirdm

Differential Revision: D44688586

fbshipit-source-id: e33b4d10e972d556cdc531504b4d66ee36c471c9