Add fields to commit message
Summary:
Change commit info view to render individual fields rather than just one big text box.
Well, at least internally there are differnet fields. In OSS, we don't know what format to use so its still just one big text box.

The idea is that we have a config that defines what fields there are, and what kind. Some are expandable text areas for lots of text, others are one-line fields which will need tokenization and autocomplete etc.

With the fields config, we can render a generic `<CommitInfoViewField>` for each which handles all cases.

Most of the heavy lifting is done by the CommitMessageFieldsUtils (which differs between internal and OSS).

The bulk of this diff is migrating things that previously hardcoded `'title'` and `'description'` fields to now use arbitrary fields.

Note that the design is not final in this diff. This is just rough draft to get something resembling fields. We'll probably do some work to better handle fields with lots of content and better information density for other fields, etc.

{F929608384}

Reviewed By: muirdm

Differential Revision: D44598445

fbshipit-source-id: 195dcc81d7cd6779fd47aa140e6d0c0e94493c48